Moved context menu to header and ran prettier
