+++ src/components/credential/CredentialCardExpanded.tsx
@@ -1,7 +1,6 @@
 import styled, {css} from '@emotion/native';
-import {useNavigation} from '@react-navigation/core';
-import React, {useContext} from 'react';
-import {Alert, Dimensions, View} from 'react-native';
+import React from 'react';
+import {Dimensions, View} from 'react-native';
 import QRCode from 'react-native-qrcode-svg';
 import {theme} from '../../../App';
 import {
@@ -10,11 +9,8 @@ import {
   vaccinationStatusColor,
   vaccinationStatusText,
 } from '../../assets/styles';
-import {DispatchAction} from '../../Reducer';
-import {Context} from '../../Store';
 import {CredentialHelper} from '../../utils/credhelper';
 import {formatAsIssuedDate} from '../../utils/date';
-import {CrendentialContextMenu} from './CredentialContextMenu';
 
 const {width} = Dimensions.get('window');
 
@@ -67,60 +63,12 @@ const topPadding = css`
   padding-top: 8px;
 `;
 
-const leftPadding = css`
-  padding-left: 48px;
-`;
-
 const bottomRadius = css`
   border-bottom-start-radius: 16px;
   border-bottom-end-radius: 16px;
 `;
 
-const reverseRow = css`
-  flex-direction: row-reverse;
-`;
-
-const columnGrow = css`
-  flex-direction: column;
-  flex-grow: 1;
-`;
-
 export const CredentialCardExpanded: React.FC<any> = ({credential}) => {
-  const [, dispatch] = useContext(Context);
-  const navigation = useNavigation();
-
-  const deleteCard = () => {
-    try {
-      dispatch({
-        type: DispatchAction.RemoveCredential,
-        payload: [credential],
-      });
-      navigation.goBack();
-    } catch (error) {
-      console.error(error);
-      Alert.alert(
-        'Yikes!',
-        'There was a problem removing this card.',
-        [{text: 'OK'}],
-        {cancelable: true},
-      );
-    }
-  };
-
-  const showCardDetails = () => {
-    try {
-      navigation.navigate('CredentialDetail' as never, {credential} as never);
-    } catch (error) {
-      console.error(error);
-      Alert.alert(
-        'Yikes!',
-        'There was a problem viewing this card.',
-        [{text: 'OK'}],
-        {cancelable: true},
-      );
-    }
-  };
-
   return (
     <View style={[{width}]}>
       <StatusView
@@ -132,27 +80,21 @@ export const CredentialCardExpanded: React.FC<any> = ({credential}) => {
             ),
           },
           bottomRadius,
-        ]}>
-        <View style={[reverseRow]}>
-          <CrendentialContextMenu
-            onDeleteTouched={deleteCard}
-            onShowDetailsTouched={showCardDetails}
-          />
-          <View style={[columnGrow, leftPadding]}>
-            <LargeText>
-              {CredentialHelper.familyNameForCredential(
-                CredentialHelper.nameForCredential(credential.record),
-              )?.toUpperCase() || ' '}
-              ,
-            </LargeText>
-            <LargeText>
-              {CredentialHelper.givenNameForCredential(
-                CredentialHelper.nameForCredential(credential.record),
-              )?.toUpperCase() || ' '}
-            </LargeText>
-          </View>
+        ]}
+      >
+        <View>
+          <LargeText>
+            {CredentialHelper.familyNameForCredential(
+              CredentialHelper.nameForCredential(credential.record),
+            )?.toUpperCase() || ' '}
+            ,
+          </LargeText>
+          <LargeText>
+            {CredentialHelper.givenNameForCredential(
+              CredentialHelper.nameForCredential(credential.record),
+            )?.toUpperCase() || ' '}
+          </LargeText>
         </View>
-
         <LargeBoldText style={[verticalPadding]}>
           {vaccinationStatusText(
             CredentialHelper.immunizationStatus(credential.record),
